==> Analysis/01_filter_11v11.ipynb <==
import os
import shutil

exclude_matches = []

ROOT = os.getcwd()                                # returns current working directory of a process.
game_data = (r"SUCCESS_SCORES")                   # creates a raw string which treats backslash (\) as a literal character. 
           
for dirMD in os.listdir(game_data):
    locMD = os.path.join(game_data, dirMD)
    for dirM in os.listdir(locMD):
        locM = os.path.join(locMD, dirM)
        if len(os.listdir(locM)) != 12:
            print(locM)                          # to check which games will be excluded
            exclude_matches.append(dirM)
            shutil.rmtree(locM)                   # to delete every games´ folder with less then 12 files     

import pandas as pd

GO = pd.read_excel('GameOverview2.xlsx')
GO

red_card = []
for row in range(0, len(GO)):
    if GO.Red_card[row]:
        # print(GO.MI_Name[row])
        red_card.append(GO.MI_Name[row])

exclude_matches.sort()
# exclude_matches

red_card.sort()
#red_card

for match in red_card:
    if match not in exclude_matches:
        print(match)
        
# DFL-MAT-00278B - no event data - incorrect should be excluded
# DFL-MAT-00278N - red card redacted - correct
# DFL-MAT-00278T - red card
# DFL-MAT-00279D - red card 
# DFL-MAT-0027C1
# DFL-MAT-0027C4
# DFL-MAT-0027CP
# DFL-MAT-0027CZ
# DFL-MAT-0027DG
# DFL-MAT-0027E8
# DFL-MAT-0027EA
# DFL-MAT-0027ES
# DFL-MAT-0027FF
# DFL-MAT-0027FJ
# DFL-MAT-0027FQ
# DFL-MAT-0027FV

shutil.rmtree('SUCCESS_SCORES/Matchday2/DFL-MAT-00278B')



for match in exclude_matches:
    if match not in red_card:
        print(match)
        
# these are just the games where we are missing the position data!
# So that's all correct!

import pickle
with open('Pickles/Games_Included.pkl', 'rb') as dic:
    games_inclued = pickle.load(dic)
    
print(len(games_inclued))
games_inclued
