==> Analysis/02_txt_to_csv.ipynb <==
# Step 1-3 combined
**Don't run the code if you have run it before!!!** It will mess up your files as it creates the files from created files!

import fileinput
import os
import csv
import numpy as np

ROOT = os.getcwd() # returns current working directory of a process.
game_data = (r"SUCCESS_SCORES") # creates a raw string which treats backslash (\) as a literal character. 
# Probably relevant for directories
type(game_data)

# MD = match day
# M = match
datasets_all = []
for dirMD in os.listdir(game_data):
    locMD = os.path.join(game_data, dirMD)
    for dirM in os.listdir(locMD):
        locM = os.path.join(locMD, dirM)
        for file in os.listdir(locM):
            datasets_all.append(file[-8:])               # creating a list of all datasets (area & interval)
            with fileinput.FileInput(os.path.join(locM, file), inplace=True, backup='.bak') as f:   
                for line in f:
                    print(line.replace(":", " "), end='')
for dirMD in os.listdir(game_data):
    locMD = os.path.join(game_data, dirMD)
    for dirM in os.listdir(locMD):
        locM = os.path.join(locMD, dirM)
        # os.chdir(locM)
        for file in os.listdir(locM):
            if file[-4:] == ".bak":
                path = os.path.join(locM, file)
                os.remove(path)
# Pause program here to check files and delete backup files if all looks good
ROOT = os.getcwd()
game_data = (r"SUCCESS_SCORES")
for dirMD in os.listdir(game_data):
    locMD = os.path.join(game_data, dirMD)
    for dirM in os.listdir(locMD):
        locM = os.path.join(locMD, dirM)
        # os.chdir(locM)
        for file in os.listdir(locM):
            with open(os.path.join(locM, file), "r") as in_txt:               # open the text file
                reader = csv.reader(in_txt, delimiter = ' ')                 # read the text file
                with open(os.path.join(locM, file)+".csv", "w") as new_csv:   # open a newly created csv
                    writer = csv.writer(new_csv)                             # write in the csv
                    for row in reader:                                       # for every row in the text file
                        #print(row)                                          # just for testing
                        new_row = []                                         # write a new row
                        for i in range(len(row)):       
                            try:
                                new_row.append(float(row[i]))                # append every element from row in text file to new_row
                            except:
                                pass
                        writer.writerow(new_row)                             # write new_row as row in csv
                        #print(new_row)
            with open(os.path.join(locM, file)+".csv", "r") as pre:           # open the cretaed csv file
                reader = csv.reader(pre, delimiter=',')                      # read the csv file
                with open(os.path.join(locM, file)+"postprocessed.csv", "w") as post:    # open a newly created csv
                    writer = csv.writer(post)                                # write in the new csv
                    for row in reader:
                        try:
                            row[0]=int(float(row[0]))                        # removing decimal places 
                            row[1]=int(float(row[1]))                        # for minute, second and ball contacts 
                            row[2]=int(float(row[2]))                        # by replacing the first, second, third and fifth 
                            row[4]=int(float(row[4]))                        # element in each row by its integer
                            writer.writerow(row)                             # write that changed row into the csv
                        except:
                            pass

datasets = list(np.unique(datasets_all))                # a list with unique datasets names
# def last4(dataset):                                     # function to sort by the last 4 digits in string (interval)
#     return dataset[-4:]
# datasets.sort(key=last4)

import pickle

with open("Pickles/Datasets.pkl", "wb") as l:
    pickle.dump(datasets, l)



datasets
