==> Analysis/06_ChiSquare.ipynb <==
import pandas as pd
import pickle
import numpy as np
from scipy.stats import chisquare

goals80 = pd.read_csv('Results/goal_Count_80_e100.csv', index_col=0)
goals80
# goals over
# data points over
# goals under
# data points under
# all datapoints

goals_over = []
goals_under = []

chi_square =[]
p_value = []

rel_over = []
rel_under = []

for dataset in goals80.columns.to_list():
    ngoals = round(goals80[dataset][0] + goals80[dataset][2],0)
    observed = []
    expected = []

    observed.append(goals80[dataset][0]) # goals over
    goals_over.append(goals80[dataset][0])
    observed.append(goals80[dataset][2]) # goals under
    goals_under.append(goals80[dataset][2])

    expected.append(goals80[dataset][1]/goals80[dataset][4]*ngoals) # expected goals over = # data points over / # all datapoint * # goals
    expected.append(goals80[dataset][3]/goals80[dataset][4]*ngoals) # expected goals under


    chi, p = chisquare(observed, expected)
    chi_square.append(chi)
    p_value.append(p)
    
    rel_over.append(goals80[dataset][0]/goals80[dataset][1])
    rel_under.append(goals80[dataset][2]/goals80[dataset][3])

chi_square_tests = pd.DataFrame(zip(goals_over, goals_under,chi_square, p_value, rel_over, rel_under),
                                columns = ['Goals_over', 'Goals_under', 'chi_square_statisitcs', 'p_values', 'rel_over', 'rel_under'])

chi_square_tests.to_csv('Results/ChiSquare_e100.csv')
chi_square_tests






ngoals = round(goals80['16m_100s'][0] + goals80['16m_100s'][2],0)
observed = []
expected = []

observed.append(goals80['16m_100s'][0]) # goals over
observed.append(goals80['16m_100s'][2]) # goals under

expected.append(goals80['16m_100s'][1]/goals80['16m_100s'][4]*ngoals) # expected goals over = # data points over / # all datapoint * # goals
expected.append(goals80['16m_100s'][3]/goals80['16m_100s'][4]*ngoals) # expected goals under


chisquare(observed, expected)

observed

expected
