==> Analysis/06_Kendall.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
from scipy import stats
import pandas as pd 

data = pd.read_csv('Results/goal_Count_8_e100.csv', index_col=0)

datasets = list(data.columns)
x = range(1,9,1)
i = 0
ds = []
ts =[]
ps= []
for dataset in datasets:
    t, p = stats.kendalltau(x, data[dataset], alternative='greater')
    print(dataset, t, p)
    ds.append(dataset)
    ts.append(t)
    ps.append(p)
    
Kendall = pd.DataFrame(list(zip(ds, ts, ps)),
               columns =['Name', 'tau', 'p-value'])

Kendall.to_csv('Results/Kendall_e100.csv')
