==> Analysis/06_MLMprep.ipynb <==
import pandas as pd
import numpy as np
import pickle
from statistics import mean


with open('Pickles/AllDatapoints_e100.pkl', 'rb') as f:
    full_data = pickle.load(f)
with open('Pickles/Datasets.pkl', 'rb') as f:
    datasets = pickle.load(f)

full_data

import os
games_included = []

game_data = (r'SUCCESS_SCORES')

for dirMD in os.listdir(game_data):
        locMD = os.path.join(game_data, dirMD)
        for dirM in os.listdir(locMD):
            games_included.append(dirM)
            
with open('Games_Included.pkl', 'wb') as l:
    pickle.dump(games_included, l)

GO = pd.read_csv('Output/GameOverview3.csv', index_col=0)
GO

teams = GO.Home.unique()
teams

team_to_games = {}

for team in teams:
    game_list = GO.MI_Name.loc[(GO.Home == team) | (GO.Away == team)].to_list() 
    team_to_games[team] =  []

    for game in game_list:
        if game in games_included: # only the games were data is included!
            team_to_games[team].append(game)  

team_to_games

games = GO.MI_Name.unique()
games

game_to_teams = {}

for game in games_included:   # only the games where i have data
    game_to_teams[game] = [GO['Home'].loc[GO.MI_Name == game].to_list()[0],GO['Away'].loc[GO.MI_Name == game].to_list()[0]] 
    
game_to_teams

# collect success scores for each team in each dataset
team_data = {}

for dataset in datasets:
    team_data[dataset] = {}
    for team in teams:
        team_data[dataset][team] = []


for dataset in datasets:
    for frame in full_data[dataset]:
        # extract match 
        match = frame[2][5:19]
        if frame[1] == -3:
            team_data[dataset][game_to_teams[match][0]].append(frame)
        elif frame[1] == -1:
            team_data[dataset][game_to_teams[match][1]].append(frame)

team_data.keys()
for key in team_data.keys():
    summ = 0
    for team in teams:
        #print(len(team_data[key][team]))
        summ = summ + len(team_data[key][team])
    print(summ)

with open('Output/TeamData_e100.pkl', 'wb') as cdict:
    pickle.dump(team_data, cdict)

# only the success scores for each team in each dataset
team_Success_Scores = {}

for dataset in datasets:
    team_Success_Scores[dataset] = {}
    for team in teams:
        team_Success_Scores[dataset][team] = []
        for frame in team_data[dataset][team]:
            team_Success_Scores[dataset][team].append(frame[0])
            
for key in team_Success_Scores.keys():
    summ = 0
    for team in teams:
        #print(len(team_data[key][team]))
        summ = summ + len(team_Success_Scores[key][team])
    print(summ)

with open('Output/team_Success_Scores_e100.pkl', 'wb') as cdict:
    pickle.dump(team_Success_Scores, cdict)

# overall average success score per team per dataset
team_Success_Score_av = {}

for dataset in datasets:
    team_Success_Score_av[dataset] = {}
    for team in teams:
        team_Success_Score_av[dataset][team] = mean(team_Success_Scores[dataset][team])
        
team_Success_Score_av

with open('Output/Team_Success_Score_av_e100.pkl', 'wb') as cdict:
    pickle.dump(team_Success_Score_av, cdict)

# data for each match of each team in each dataset
team_game_data = {}

for dataset in datasets:
    team_game_data[dataset] = {}
    for team in teams:
        team_game_data[dataset][team] = {}
        for match in team_to_games[team]:
            team_game_data[dataset][team][match] = []
            for frame in team_data[dataset][team]:
                if match in frame[2]:
                    team_game_data[dataset][team][match].append(frame)

team_game_data['16m_100s']['FC Bayern München']

with open('Output/team_game_data_e100.pkl', 'wb') as cdict:
    pickle.dump(team_game_data, cdict)

# success for each match of each team in each dataset
team_Success_Score_game = {}

for dataset in datasets:
    team_Success_Score_game[dataset] = {}
    for team in teams:
        team_Success_Score_game[dataset][team] = {}
        for match in team_to_games[team]:
            team_Success_Score_game[dataset][team][match] = []
            for frame in team_game_data[dataset][team][match]:
                team_Success_Score_game[dataset][team][match].append(frame[0])

team_Success_Score_game['16m_100s']['1. FSV Mainz 05']

with open('Output/team_Success_Score_game_e100.pkl', 'wb') as cdict:
    pickle.dump(team_Success_Score_game, cdict)

team_Success_Score_game_av = {}

for dataset in datasets:
    team_Success_Score_game_av[dataset] = {}
    for team in teams:
        team_Success_Score_game_av[dataset][team] = {}
        for match in team_to_games[team]:
            team_Success_Score_game_av[dataset][team][match] = mean(team_Success_Score_game[dataset][team][match])
            
            


team_Success_Score_game_av

with open('Output/team_Success_Score_game_av_e100.pkl', 'wb') as cdict:
    pickle.dump(team_Success_Score_game_av, cdict)

Quality_dict = {}
Quality_dict['Points'] = {
    'FC Bayern München' : 82, 
    'SV Werder Bremen' : 45,
    'Borussia Dortmund' : 64,
    '1. FSV Mainz 05': 67,
    'Borussia Mönchengladbach' : 45,
    'Bayer 04 Leverkusen' : 41,
    'Hertha BSC' : 49,
    'Sport-Club Freiburg' : 48,
    '1. FC Köln' : 49,
    'SV Darmstadt 98' : 25,
    'Hamburger SV' : 38,
    'FC Ingolstadt 04' : 32,
    'FC Augsburg' : 38,
    'VfL Wolfsburg' : 37,
    'TSG 1899 Hoffenheim' : 62,
    'RB Leipzig' : 67,
    'FC Schalke 04' : 43,
    'Eintracht Frankfurt' : 42 
}
Quality_dict['Ranking'] = {
    'FC Bayern München' : 1, 
    'SV Werder Bremen' : 8,
    'Borussia Dortmund' : 3,
    '1. FSV Mainz 05': 15,
    'Borussia Mönchengladbach' : 9,
    'Bayer 04 Leverkusen' : 12,
    'Hertha BSC' : 6,
    'Sport-Club Freiburg' : 7,
    '1. FC Köln' : 5,
    'SV Darmstadt 98' : 18,
    'Hamburger SV' : 14,
    'FC Ingolstadt 04' : 17,
    'FC Augsburg' : 13,
    'VfL Wolfsburg' : 16,
    'TSG 1899 Hoffenheim' : 4,
    'RB Leipzig' : 2,
    'FC Schalke 04' : 10,
    'Eintracht Frankfurt' : 11
}

Quality_dict['Goal_difference'] = {
    'FC Bayern München' : 67, 
    'SV Werder Bremen' : -3,
    'Borussia Dortmund' : 32,
    '1. FSV Mainz 05': -11,
    'Borussia Mönchengladbach' : -4,
    'Bayer 04 Leverkusen' : -2,
    'Hertha BSC' : -4,
    'Sport-Club Freiburg' : -18,
    '1. FC Köln' : 9,
    'SV Darmstadt 98' : -35,
    'Hamburger SV' : -28,
    'FC Ingolstadt 04' : -21,
    'FC Augsburg' : -16,
    'VfL Wolfsburg' : -18,
    'TSG 1899 Hoffenheim' : 27,
    'RB Leipzig' : 27,
    'FC Schalke 04' : 5,
    'Eintracht Frankfurt' : -7
}

with open('Pickles/Quality_dict.pkl', 'wb') as dic:
    pickle.dump(Quality_dict, dic)
Quality_dict

with open('Output/team_Success_Score_game_av.pkl', 'rb') as cdict:
    team_Success_Score_game_av = pickle.load(cdict)
    
with open('Output/team_Success_Score_game.pkl', 'rb') as cdict:
    team_Success_Score_game = pickle.load(cdict)
    
with open('Games_Included.pkl', 'rb') as l:
    games_included = pickle.load(l)
    
GO = pd.read_csv('Output/GameOverview3.csv', index_col=0)


# table for mixed model

# game
# team (--> TQ)
# opponent --> TQ
# home or away (dummy)
# possession percentage
# match outcome (categorical)
# match outcome (difference)
# average Success Score (ASS) 16m100s for that game
# TQ_points, ranking & goal difference
# OQ see above
# ASS 16m300s
# ASS 16m500s
# ASS 30m100s
# ASS 30m300s
# ASS 30m500s

# outcome dictionary to match team outcome to opponents logical outcome
outcomes = {1: -1,
            0: 0,
           -1: 1}

game_column = []
team_column = []
opponent_column = []
location_column = []
possession_column = []
outcome_cat_column = []
outcome_num_column = []
TQ_r = []
TQ_p = []
TQ_gd = []
OQ_r = []
OQ_p = []
OQ_gd = []
ASS_16m_100s = []
ASS_16m_300s = []
ASS_16m_500s = []
ASS_30m_100s = []
ASS_30m_300s = []
ASS_30m_500s = []
ASS_16m_100s_diff = []
ASS_16m_300s_diff = []
ASS_16m_500s_diff = []
ASS_30m_100s_diff = []
ASS_30m_300s_diff = []
ASS_30m_500s_diff = []
ASS_16m_100s_diff_per = []
ASS_16m_300s_diff_per = []
ASS_16m_500s_diff_per = []
ASS_30m_100s_diff_per = []
ASS_30m_300s_diff_per = []
ASS_30m_500s_diff_per = []

for game in games_included:
    game_column.append(game)
    game_column.append(game)
    location_column.append(1) # 1 = home
    location_column.append(0) # 0 = away
    Home = GO.Home[GO.MI_Name == game].to_list()[0] # Home team = Team in 1 and opponent in 2
    Away = GO.Away[GO.MI_Name == game].to_list()[0] # Away team = Team in 2 and opponent in 1
    team_column.append(Home)
    team_column.append(Away)
    possession_column.append(GO.Pos_Home[GO.MI_Name == game].to_list()[0])
    possession_column.append(GO.Pos_Away[GO.MI_Name == game].to_list()[0])
    opponent_column.append(Away)
    opponent_column.append(Home)
    
    # determine categorical outcome (W, D, L)
    goals_scored = GO.Final_Score[GO.MI_Name == game].to_list()[0].split(':')
    if int(goals_scored[0]) >  int(goals_scored[1]):
        Home_Outcome = 1
    elif  int(goals_scored[0]) <  int(goals_scored[1]):
        Home_Outcome = -1
    else:
        Home_Outcome = 0
    Away_Outcome = outcomes[Home_Outcome] # match opponents outcome 
    
    outcome_cat_column.append(Home_Outcome)
    outcome_cat_column.append(Away_Outcome)
    
    # determine difference in goals scored
    Home_Outcome = int(goals_scored[0]) - int(goals_scored[1])
    Away_Outcome = Home_Outcome *-1
    
    outcome_num_column.append(Home_Outcome)
    outcome_num_column.append(Away_Outcome)
    
    # TQ = team quality
    TQ_r.append(Quality_dict['Ranking'][Home])
    TQ_p.append(Quality_dict['Points'][Home])
    TQ_gd.append(Quality_dict['Goal_difference'][Home])
    TQ_r.append(Quality_dict['Ranking'][Away])
    TQ_p.append(Quality_dict['Points'][Away])
    TQ_gd.append(Quality_dict['Goal_difference'][Away])
    
    # OQ = opponent team quality
    OQ_r.append(Quality_dict['Ranking'][Away])
    OQ_p.append(Quality_dict['Points'][Away])
    OQ_gd.append(Quality_dict['Goal_difference'][Away])
    OQ_r.append(Quality_dict['Ranking'][Home])
    OQ_p.append(Quality_dict['Points'][Home])
    OQ_gd.append(Quality_dict['Goal_difference'][Home])
    
    # success scores for each game
    SSH = team_Success_Score_game_av['16m_100s'][Home][game]
    SSA = team_Success_Score_game_av['16m_100s'][Away][game]
    ASS_16m_100s.append(SSH)
    ASS_16m_100s.append(SSA)
    ASS_16m_100s_diff.append(SSH - SSA)
    ASS_16m_100s_diff.append(SSA - SSH)
    
    
    SSH = team_Success_Score_game_av['16m_300s'][Home][game]
    SSA = team_Success_Score_game_av['16m_300s'][Away][game]
    ASS_16m_300s.append(SSH)
    ASS_16m_300s.append(SSA)
    ASS_16m_300s_diff.append(SSH - SSA)
    ASS_16m_300s_diff.append(SSA - SSH)
    
    SSH = team_Success_Score_game_av['16m_500s'][Home][game]
    SSA = team_Success_Score_game_av['16m_500s'][Away][game]
    ASS_16m_500s.append(SSH)
    ASS_16m_500s.append(SSA)    
    ASS_16m_500s_diff.append(SSH - SSA)
    ASS_16m_500s_diff.append(SSA - SSH)
    
    
    SSH = team_Success_Score_game_av['30m_100s'][Home][game]
    SSA = team_Success_Score_game_av['30m_100s'][Away][game]
    ASS_30m_100s.append(SSH)
    ASS_30m_100s.append(SSA)
    ASS_30m_100s_diff.append(SSH - SSA)
    ASS_30m_100s_diff.append(SSA - SSH)
    
    SSH = team_Success_Score_game_av['30m_300s'][Home][game]
    SSA = team_Success_Score_game_av['30m_100s'][Away][game]
    ASS_30m_300s.append(SSH)
    ASS_30m_300s.append(SSA)
    ASS_30m_300s_diff.append(SSH - SSA)
    ASS_30m_300s_diff.append(SSA - SSH)
    
    SSH = team_Success_Score_game_av['30m_500s'][Home][game]
    SSA = team_Success_Score_game_av['30m_500s'][Away][game]
    ASS_30m_500s.append(SSH)
    ASS_30m_500s.append(SSA)
    ASS_30m_500s_diff.append(SSH - SSA)
    ASS_30m_500s_diff.append(SSA - SSH)
    
    
MLM = pd.DataFrame(zip(game_column, team_column, location_column, outcome_cat_column, outcome_num_column,possession_column,
                       TQ_r, TQ_p, TQ_gd, OQ_r, OQ_p, OQ_gd,
                       ASS_16m_100s, ASS_16m_300s, ASS_16m_500s, ASS_30m_100s, ASS_30m_300s, ASS_30m_500s,
                      ASS_16m_100s_diff, ASS_16m_300s_diff, ASS_16m_500s_diff, ASS_30m_100s_diff, ASS_30m_300s_diff, ASS_30m_500s_diff),
                  columns = ['Game', 'Team', 'Location', 'Outcome_cat', 'Outcome_num', 'possession',
                             'TQ_r', 'TQ_p', 'TQ_gd', 'OQ_r', 'OQ_p', 'OQ_gd',
                            'ASS_16m_100s', 'ASS_16m_300s', 'ASS_16m_500s', 'ASS_30m_100s', 'ASS_30m_300s', 'ASS_30m_500s',
                            'ASS_16m_100s_diff', 'ASS_16m_300s_diff', 'ASS_16m_500s_diff', 'ASS_30m_100s_diff', 'ASS_30m_300s_diff', 'ASS_30m_500s_diff'])

MLM

MLM.to_csv('Results/MLM_data_e100.csv')

import pandas as pd
MLM = pd.read_csv('Results/MLM_data_e100.csv')
MLM

import pandas as pd
import numpy as np
import pickle
from statistics import mean
from scipy.stats import iqr
scores = MLM.columns.to_list()[-6:]

N = []
mean = []
median = []
sd = []
P25 = []
P75 = []
IQR = []
Min = []
Max = []

for s in scores:
    data = MLM[s][MLM[s] > 0]
    N.append(len(data))
    mean.append(data.mean())
    median.append(data.median())
    sd.append(data.std())
    Min.append(min(data))
    Max.append(max(data))
    P25.append(np.percentile(data, 25))
    P75.append(np.percentile(data, 75))
    IQR.append(iqr(data))
    

game_desc = pd.DataFrame(zip(N, mean, median, sd, Min, Max, P25, P75, IQR),
                         columns=['N', 'Mean', 'Median', 'SD', 'Min', 'Max', 'P25', 'P75', 'IQR'],
                        index = scores)

game_desc

game_desc.to_csv('Results/game_desc.csv')

import matplotlib.pyplot as plt
from matplotlib import colors
from matplotlib.ticker import PercentFormatter

fig, axs = plt.subplots(2, 3, sharey=False, tight_layout=True)

for s, score in enumerate(scores):
    y = 0 if s <= 2 else 1
    x = s if y == 0 else s-3
    # print(s, x, y)
    lim =  10 if s <= 2 else 20
    stepsize =  lim/2
    axs[y,x].hist(MLM[score][MLM[score] > 0], density = True, color = '#07607e',
                  bins = np.arange(0, max(MLM[score]), 0.5))
    axs[y,x].set_xlim([0, lim])
    axs[y,x].xaxis.set_ticks(np.arange(0, lim+1, stepsize))
    axs[y,x].axvline(x = game_desc.P25[s], ymin = 0, ymax= 0.98, color = '#f13b38', linestyle = 'dashed')
    axs[y,x].axvline(x = game_desc.P75[s], ymin = 0, ymax= 0.98, color = '#f13b38', linestyle = 'dashed')

    #axs[y,x].spines['top'].set_visible(False)
    #axs[y,x].spines['right'].set_visible(False)
    axs[y,x].spines['top'].set_color('#B1B1B1')
    axs[y,x].spines['right'].set_color('#B1B1B1')
    axs[y,x].spines['left'].set_color('#B1B1B1')
    axs[y,x].spines['bottom'].set_color('#B1B1B1')



import matplotlib
matplotlib.pyplot.close(fig = 'all')

groupingvar1 = ['16m', '30m']
groupingvar2 = ['100s', '300s', '500s']
colors = ['#5D8DAF', '#8daf5d']
# colors = ['#5D8DAF', '#f24e4b']

with plt.style.context('C:/Users/DavidB/PythonProjects/Style/StyleScientific.mplstyle'):
    fig, axes = plt.subplots(2,3, sharey=False, layout='constrained')
    for s, score in enumerate(scores):
        data = MLM[score][MLM[score] > 0]
        y = 0 if s <= 2 else 1
        x = s if y == 0 else s-3       
        limx =  10 if s <= 2 else 20
        stepsizex =  limx/2
        
        limy = 0.5 if s <= 2 else 0.2
        stepsizey = limy/5
        
        axes[y, x].hist(data, density = True, color = colors[y], bins = np.arange(0, max(MLM[score]), 0.5),
                            label = ''.join([groupingvar1[y], groupingvar2[x]]))
        axes[y,x].axvline(x = game_desc.P25[s], ymin = 0, ymax= 0.98, color = '#f13b38', linestyle = 'dashed')
        axes[y,x].axvline(x = game_desc.P75[s], ymin = 0, ymax= 0.98, color = '#f13b38', linestyle = 'dashed')
        axes[y,x].set_xlim([0, limx])
        axes[y,x].xaxis.set_ticks(np.arange(0, limx+1, stepsizex))
        axes[y,x].set_ylim([0, limy])
        axes[y,x].yaxis.set_ticks(np.arange(0, limy+0.01, stepsizey))
        axes[y,x].tick_params(axis='both', labelsize = 8)
        
        if x == 0:
            axes[y,x].set_ylabel(groupingvar1[y], fontsize = 9)
        if y == 1:
            axes[y,x].set_xlabel(groupingvar2[x], fontsize = 9)
        
    fig.supxlabel('Success-Score', size =10)
    fig.supylabel('Density', size = 10)
    #fig.legend(title= 'Success-Score', loc='outside right upper', fontsize = 8, title_fontsize = 9)


fig
